requirements.txt
google-api-python-client
wikiapi
flask
wordcloud
